+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 * Fix rxjs imports
 * Add optional hd parameter for Google Auth with particular G Suite domain, see https://developers.google.com/identity/protocols/OpenIDConnect#hd-param
 * fix: local_state is always null because is not being set
+* fix: change for emtpy header in id_token, improved logging
 
 <a name="2017-10-05"></a>
 ### 2017-10-05 version 1.3.18